Add shared Venice service client
