+++ backend/src/services/venice/index.ts
@@ -1,0 +1,9 @@
+export { VeniceClient } from './client.js';
+export type {
+  AgentType,
+  CompleteOptions,
+  VeniceChatOptions,
+  VeniceClientConfig,
+  VeniceClientLike,
+  VeniceMessage,
+} from './types.js';
